+++ vignettes/sp-df.R
@@ -7,9 +7,9 @@ library(maptools)
 library(dplyr)
 data(wrld_simpl)
 
-## ------------------------------------------------------------------------
-library(ggplot2)
-x <- bind_cols(as.data.frame(wrld_simpl), as_data_frame(as.data.frame(raster::geom(wrld_simpl))) %>% group_by(cump, object) %>% nest %>%  group_by(object) %>% nest)
+## ----eval=FALSE----------------------------------------------------------
+#  library(ggplot2)
+#  x <- bind_cols(as.data.frame(wrld_simpl), as_data_frame(as.data.frame(raster::geom(wrld_simpl))) %>% group_by(cump, object) %>% nest %>%  group_by(object) %>% nest)
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  vertices <- function(x) {